Extract mapErrors to statusMixin
